Support newer zookeeper packaging
Zookeeper in the official 3.5.6 releases moved to separating out the
binary and source into two different archives.  On top of that all
packages are prepended with `apache-` now.

This changes the make file to look for the minor version and use a
different package name.

Also added a `clean` Makefile target to remove this stuff